expense-tracker/backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
